Make MiniMaven more consistent with the imagej-maven-plugin
When the imagej.app.directory property is set, the imagej-maven-plugin
creates the directory if necessary.

While at it, restructure the source code to reflect the fact that we do
not necessarily require a project with "jar" packaging, by renaming the
API's function to buildAndInstall().

Also, offer a method to install to a given directory without setting a
system property to the path of said directory first.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>