MiniMaven: add a Class-Path entry to the manifest
The .jar file format specification mandates that regular .jar
files' manifest can contain a 'Class-Path' entry that contains a
space-delimited list of class path entries (relative to the .jar file
containing the manifest) that should be appended to the class path
by the ClassLoader implicitly.

This useful feature allows adding a .jar file to the class path and use
its classes without having to add transitive dependencies on the class
path explicitly.

Example: Let's assume that MiniMaven is contained in a .jar file called
ij-minimaven-2.0.0-SNAPSHOT.jar and it uses classes included in the
dependency ij-core-2.0.0-SNAPSHOT.jar. Adding the appropriate Class-Path
entries to the manifest allows executing the main class using

	java -jar ij-minimaven-2.0.0.jar

without having to specify the dependency ij-core-2.0.0-SNAPSHOT.jar
explicitly. Indeed, if ij-core is upgraded to a version that requires,
say, scijava-common and specified that dependency in its own manifest,
launching MiniMaven using the same command-line will still work,
provided that all dependencies can be found in the specified (relative)
paths.

The Class-Path entry is expected to work at least with URLClassLoaders,
at least since Java 1.3.

Of course, Maven adds the Class-Path entry to the manifest already, so
let's do the same thing with MiniMaven.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>