Refactor findVersion()
The findVersion() method is supposed to traverse the
<dependencyManagement> and also the parents' <dependencies> and
<dependencyManagement> sections to find the appropriate <version> tag
for dependencies whose version is not set in the current project's
<dependencies> section.

But the method itself was conflating two concepts: querying the
dependencyManagement information, and returning just one version.

For exclusions, we require the first concept: to query the
dependencyManagement information (because Bio-Formats' kryo dependency
has exclusions managed that way).

So let's refactor it.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>