Handle <exclusions> properly
We tried to punt on the issue and just filter dependencies *after*
gathering them.

That did not work, however: if a dependency has exclusions, the
*transitive* dependencies of the excluded ones *also* need to be
excluded.

So let's stop punting on the issue and do it properly.

While at it, heed exclusions declared in <dependencyManagement>
sections as well as exclusions declared in the parents.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>