Fix MavenProject#compareTo()
We would like to sort MavenProjects into TreeSets; For that, they
have to be Comparable. The previous version of the compareTo()
method failed to take classifiers into account; As a consequence,
MavenProject instances representing artifacts with classifiers were
ignored when accumulating dependencies, if the same artifact is a
dependency *without* the classifier already. This happened for example
with ImageJ2's miglayout dependency (which comes as an artifact without
a classifier as well as an artifact with the 'swing' classifier).

While at it, we also implement the equals() and hashCode() methods,
as mandated by the Comparable contract.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>