Always download to temporary files before verifying the checksum
When downloading files together with their .sha1 files, we do verify
the checksums afterwards. But we do not remove the files (so that we
can investigate what went wrong later). So we need to download said
files to temporary files and rename them to the correct file names
only after verifying that everything's groovy.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>