Use the caching URL of ImageJ's Maven mirror
ImageJ's Nexus server, responsible for serving and caching the Maven
repositories used by the various projects in the ImageJ ecosystem, has
several URLs to offer.

We used to access cached artifacts -- without looking at, say, OSS
Sonatype or Maven Central when the artifacts were not found.

Now we use the *caching* URL -- read: whenever artifacts are not found
in the cache, we let the Nexus server look actively in the upstream
repositories for our artifacts.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>