If an aggregator POM specifies explicit dependencies, install them
The sc.fiji:fiji project defines dependencies even if it does not build
an artifact itself. It is a convenient parent for lazy people who simply
want to drag in all of the Fiji components. Also, when calling the
'install' command, we actually want those dependencies installed into
the Fiji.app/ directory.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>