JavaCompiler: try ToolsProvider first
When compiling, check the javax.tools.ToolsProvider compiler first.
Failing that we can look for com.sun.tools and then fall back to the
system javac.

This should allow - for example - Fiji to move away from shipping a
com.sun.tools jar as long as it's launched with a JDK.

Closes #9